requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
decorator==4.4.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
six==1.15.0
tornado==6.0.4
tzlocal==2.1
Werkzeug==1.0.1
